// src/tagline/types.ts
export const enum Styles {
  Style,
  Size,
  Radius,
  Align,
}

export const enum Style {
  LightGrey = 'light-grey',
  DarkBlue = 'dark-blue',
  LightBlue = 'light-blue',
  Transparent = 'transparent',
}

export const enum Size {
  XL = 'xl',
  L = 'l',
  M = 'm',
  S = 's',
  XS = 'xs',
}

export const enum Radius {
  Zero = '0',
  Four = '4',
  Eight = '8',
  Twelve = '12',
  Hundred = '100'
}

export const enum Align {
  Left = 'left',
  Center = 'center',
  Right = 'right',
}